Created placeholder login page and drawer widget.
Login page only has functionalities within on the Login button, which uses Navigator.pushReplacement to bring the user to the home page. Pressing "back" in the home screen will exit the app.
To go back to the login page, a drawer is added with a logout button included. It uses Navigator.pushReplacement as well, pressing "back" in the login page after logging out will also exit the app.