Updated Scanner Page
Scanner now has 2 outputs: OCR output and translation output.
The user can edit the OCR output in the case of unnecessary noise being scanned.
A translation button has been added to translate the OCR output using the "translator" Flutter package, it utilizes Google Translate. The output will be displayed below within a Text widget.
The image display has also been updated for the user to interact with. Upon tapping the image preview, it will be then be displayed in fullscreen with zooming functionalities included.
The "Add to Dictionary" button is still a placeholder.
Inputs and outputs will be retained after moving to another widget, It will only be deleted upon exiting the app or overwritten by new inputs and outputs.